authentication added

+++ TodoApp/main.py
@@ -4,10 +4,12 @@ import models
 from sqlalchemy.orm import Session
 from pydantic import BaseModel,Field
 from typing import Optional
+from auth import get_current_user,get_user_exception
 
 app = FastAPI()
 
 models.Base.metadata.create_all(bind=engine)
+
 
 def get_db():
     try:
@@ -27,6 +29,12 @@ class Todo(BaseModel):
 async def read_all(db: Session=Depends(get_db)):
     return db.query(models.Todos).all()
 
+@app.get('/todos/user')
+async def read_all_by_user(user: dict= Depends(get_current_user),db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+    return db.query(models.Todos).filter(models.Todos.owner_id == user.get('id')).all()    
+
 @app.get('/todo/{todo_id}')
 async def read_todo(todo_id: int, db: Session = Depends(get_db)):
     todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
@@ -35,12 +43,15 @@ async def read_todo(todo_id: int, db: Session = Depends(get_db)):
     raise http_exception()
 
 @app.post('/')
-async def create_todo(todo:Todo, db: Session = Depends(get_db)):
+async def create_todo(todo:Todo,user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
     todo_model = models.Todos()
     todo_model.title = todo.title
     todo_model.description = todo.description
     todo_model.priority = todo.priority
     todo_model.complete = todo.complete
+    todo_model.owner_id = user.get('id')
 
     db.add(todo_model)
     db.commit()
@@ -48,8 +59,10 @@ async def create_todo(todo:Todo, db: Session = Depends(get_db)):
     return success_response(201)
 
 @app.put('/{todo_id}')
-async def update_todo(todo_id: int,todo:Todo, db: Session = Depends(get_db)):
-    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+async def update_todo(todo_id: int,todo:Todo,user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).filter(models.Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise http_exception()
     
@@ -64,11 +77,13 @@ async def update_todo(todo_id: int,todo:Todo, db: Session = Depends(get_db)):
     return success_response(201)
 
 @app.delete('/{todo_id}')
-async def delete_todo(todo_id: int, db: Session = Depends(get_db)):
+async def delete_todo(todo_id: int,user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
     todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
     if todo_model is None:
         raise http_exception()
-    db.query(models.Todos).filter(models.Todos.id == todo_id).delete()
+    db.query(models.Todos).filter(models.Todos.id == todo_id).filter(models.Todos.owner_id == user.get('id')).delete()
     db.commit()
 
     return success_response(201)
